Makefile: rename "test_bindir_programs" variable, pre-declare
Pre-declare the "BIN-WRAPPERS" variable, renamed from
"test_bindir_programs", which is more in line with our usual naming
conventions in the Makefile.

There's also no reason to use a "simply expanded variable" here, so
let's use a normal "=" assignment. It doesn't make a difference in the
end as we're about to use to to define a rule, so it'll have
"IMMEDIATE" expansion here.

See ea925196f1b (build dashless "bin-wrappers" directory similar to
installed bindir, 2009-12-02) for the introduction of
"test_bindir_programs".

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Taylor Blau <me@ttaylorr.com>