revision: add new parameter to specify all visible refs
Users can optionally hide refs from remote users in git-upload-pack(1),
git-receive-pack(1) and others via the `transfer.hideRefs`, but there is
not an easy way to obtain the list of all visible or hidden refs right
now. We'll require just that though for a performance improvement in our
connectivity check.

Add a new pseudo-ref `--visible-refs=` that pretends as if all refs have
been added to the command line that are not hidden. The pseudo-ref
requiers either one of "transfer", "uploadpack" or "receive" as argument
to pay attention to `transfer.hideRefs`, `uploadpack.hideRefs` or
`receive.hideRefs`, respectively.

Signed-off-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Taylor Blau <me@ttaylorr.com>